Move implementation to base_spark
