+++ featurebyte/session/databricks.py
@@ -173,17 +173,3 @@ class DatabricksSession(BaseSparkSession):
                 for record_batch in arrow_table.to_batches():
                     counter += 1
                     yield record_batch
-
-    def _format_comment(self, comment: str) -> str:
-        return self.sql_to_string(make_literal_value(comment))
-
-    async def comment_table(self, table_name: str, comment: str) -> None:
-        formatted_table = self.format_quoted_identifier(table_name)
-        query = f"COMMENT ON TABLE {formatted_table} IS {self._format_comment(comment)}"
-        await self.execute_query(query)
-
-    async def comment_column(self, table_name: str, column_name: str, comment: str) -> None:
-        formatted_table = self.format_quoted_identifier(table_name)
-        formatted_column = self.format_quoted_identifier(column_name)
-        query = f"ALTER TABLE {formatted_table} ALTER COLUMN {formatted_column} COMMENT {self._format_comment(comment)}"
-        await self.execute_query(query)
